Deploy to Pypi on tags only
When new commit is pushed, Travis
will run the tests, but will not
deploy results on Pypi. Only on new
tag Travis will push code to Pypi.
The package version is resolved on
the base of the closest tag which
contains current commit. If it's
not possible (there is not tags),
set the version on the commit's
timestamp.