Add regex in branch section of Travis
When new tag is pushed the original
`branches` section skipped all the
branches except `master`, so the new
tag (which is considered as branch)
was not built. To allow build `master`
and tags the regex filter was added.
The tags should have the names similar
to v1.0.0 or v1.0.0-beta